RemotePCIeTransport.get_system_information_from_transport

+++ scripts/casperfpga_rest_server.py
@@ -7,6 +7,7 @@ from flask_restful import Resource, Api, reqparse
 
 from io import StringIO
 import re
+import argparse
 import glob
 import os
 import base64
@@ -19,6 +20,9 @@ __author__ = 'radonnachie'
 __date__ = 'Feb 2022'
 
 LOGGER = logging.getLogger(__name__)
+# Set log format to DEBUG  level
+logging.basicConfig(format="%(asctime)s %(levelname)s %(message)s",
+        level=logging.INFO)
 
 UPLOAD_FOLDER = '/tmp'
 ALLOWED_EXTENSIONS = {'fpg'}
@@ -32,23 +36,21 @@ app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 
 TRANSPORT_TARGET_DICT = {}
 XDMA_PCIE_DICT = None
-
-def LOG_INFO(str):
-    LOGGER.info(str)
-    print(str)
+PCIE_XDMA_DICT = None
 
 class TransportTarget(object):
-    def __init__(self, target, instance_id = 0, cfpga = None):
-        self.target = target
-        self.instance_id = instance_id
+    def __init__(self, xdma_id, cfpga = None):
+        self.target = 'pcie%s'%XDMA_PCIE_DICT[xdma_id]
+        self.instance_id = int(xdma_id)
         self.cfpga = cfpga
         self.fpgfile_path = None
+        self.fpg_template = None
 
         if self.cfpga is None:
-            LOG_INFO('Connecting to "{}"'.format(target))
+            LOGGER.info('Connecting to "{}"'.format(self.target))
             self.cfpga = casperfpga.CasperFpga(
-                        host=target,
-                        instance_id=instance_id,
+                        host=self.target,
+                        instance_id=self.instance_id,
                         transport=casperfpga.LocalPcieTransport,
                 )
     
@@ -57,21 +59,44 @@ class TransportTarget(object):
             self.cfpga.disconnect()
     
     def _setFpgfile_path(self, fpgfile_path):
-        LOG_INFO('"{}" is programmed with "{}"'.format(self.target, fpgfile_path))
+        LOGGER.info('"{}" is programmed with "{}"'.format(self.target, fpgfile_path))
         self.fpgfile_path = fpgfile_path
         self.cfpga.get_system_information(self.fpgfile_path)
+        fpg_header = parse_fpg(self.fpgfile_path)[0]
+        self.fpg_template = None
+        try:
+            self.fpg_template = fpg_header[list(fpg_header.keys())[0]]['pr_template']
+        except:
+            pass
+    
+    def _fpgfileMatchesCurrentTemplate(self, fpgfile_path):
+        new_fpg_header = parse_fpg(fpgfile_path)[0]
+        new_fpg_template = None
+        try:
+            new_fpg_template = new_fpg_header[list(new_fpg_header.keys())[0]]['pr_template']
+        except:
+            pass
+        
+        if(self.fpg_template is not None and new_fpg_template != self.fpg_template):
+            LOGGER.info('Previously uploaded fpg file used template {}. Current '
+            'fpg supplied for programming uses template {}. '
+            'This is not a permitted action.'.format(self.fpg_template, new_fpg_template))
+            return False
+        return True
     
     def upload_to_ram_and_program(self, fpgfile_path):
         if fpgfile_path != self.fpgfile_path:
+            if not self._fpgfileMatchesCurrentTemplate(fpgfile_path):
+                return False
+            
             if (self.fpgfile_path is not None and
                 os.path.exists(self.fpgfile_path) and
                 self.fpgfile_path.startswith(app.config['UPLOAD_FOLDER'])
             ):
-
-                LOG_INFO('Removing previously uploaded fpg file for "{}": "{}"'.format(self.target, fpgfile_path))
+                LOGGER.info('Removing previously uploaded fpg file for "{}": "{}"'.format(self.target, fpgfile_path))
                 os.remove(self.fpgfile_path)
             
-            LOG_INFO('Programming "{}" with "{}"'.format(self.target, fpgfile_path))
+            LOGGER.info('Programming "{}" with "{}"'.format(self.target, fpgfile_path))
             self.cfpga.transport.upload_to_ram_and_program(fpgfile_path)
             self._setFpgfile_path(fpgfile_path)
         return True
@@ -79,27 +104,38 @@ class TransportTarget(object):
     def is_connected(self, timeout=None, retries=None):
         return self.cfpga.transport.is_connected(timeout, retries)
 
-def getTransportTarget(target, instance_id=0):
-    target_id = '{}.{}'.format(target, instance_id)
-    if target_id not in TRANSPORT_TARGET_DICT:
-        TRANSPORT_TARGET_DICT[target_id] = TransportTarget(target, instance_id)
-    return TRANSPORT_TARGET_DICT[target_id]
+def getXdmaIdFromTarget(target):
+    if target.startswith('pcie'):
+        pci_id = target[4:]
+        if pci_id not in PCIE_XDMA_DICT:
+            raise RuntimeError(
+                'pci_id "{}" not recognised:\n{}'.format(
+                    pci_id, PCIE_XDMA_DICT
+                )
+            )
+        return PCIE_XDMA_DICT[pci_id]
+    
+    if target.startswith('xdma'):
+        return target[4:]
 
-def getInstanceIdFromRequestArgs(args):
-    pci_id = args.get('pci_id', default = None, type = str)
-    if pci_id is not None:
-        if pci_id not in XDMA_PCIE_DICT:
-            raise RuntimeError('pci_id "{}" not recognised:\n{}'.format(pci_id, XDMA_PCIE_DICT))
-        return XDMA_PCIE_DICT[pci_id]
-    return args.get('instance_id', default = 0, type = int)
+    raise RuntimeError((
+        'Specified target "{}" not recognised:\nmust begin with either "pcie"'
+        ' or "xdma" or be an exact XDMA ID ({}).').format(target,
+        XDMA_PCIE_DICT.keys()
+    ))
+
+def getTransportTarget(target):
+    xdma_id = getXdmaIdFromTarget(target) if target not in XDMA_PCIE_DICT.keys() else target
+    if xdma_id not in TRANSPORT_TARGET_DICT:
+        TRANSPORT_TARGET_DICT[xdma_id] = TransportTarget(xdma_id)
+    return TRANSPORT_TARGET_DICT[xdma_id]
 
 class RestTransport_Device(Resource):
     def get(self, target, device_name):
         size = request.args.get('size', type = int)
         offset = request.args.get('offset', default = 0, type = int)
-
-        instance_id = getInstanceIdFromRequestArgs(request.args)        
-        transportTarget = getTransportTarget(target, instance_id)
+      
+        transportTarget = getTransportTarget(target)
 
         try:
             bytes_read = transportTarget.cfpga.transport.read(device_name, size, offset)
@@ -118,8 +154,7 @@ class RestTransport_Device(Resource):
     def put(self, target, device_name):
         offset = request.args.get('offset', default = 0, type = int)
 
-        instance_id = getInstanceIdFromRequestArgs(request.args)        
-        transportTarget = getTransportTarget(target, instance_id)
+        transportTarget = getTransportTarget(target)
         data = request.data
 
         try:
@@ -143,8 +178,7 @@ class RestTransport_DeviceList(Resource):
     def get(self, target):
         device_name = request.args.get('device_name', default = None, type = str)
 
-        instance_id = getInstanceIdFromRequestArgs(request.args)        
-        transportTarget = getTransportTarget(target, instance_id)
+        transportTarget = getTransportTarget(target)
 
         try:
             response_val = list(transportTarget.cfpga.listdev())
@@ -163,23 +197,21 @@ class RestTransport_DeviceList(Resource):
 
 class RestTransport_FpgFile(Resource):
     def get(self, target):
-        instance_id = getInstanceIdFromRequestArgs(request.args)        
-        transportTarget = getTransportTarget(target, instance_id)
+        transportTarget = getTransportTarget(target)
 
         try:
-            send_file(transportTarget.fpgfile_path)
+            return send_file(transportTarget.fpgfile_path)
         except:
             response = {
                 'response':  'Failed to send fpgafile: {}'.format(transportTarget.fpgfile_path)
             }
             code = 500
-        return response, code
+            return response, code
 
 
     def put(self, target):
         filepath = None
-        instance_id = getInstanceIdFromRequestArgs(request.args)        
-        transportTarget = getTransportTarget(target, instance_id)
+        transportTarget = getTransportTarget(target)
 
         try:
             if 'fpga' in request.files:
@@ -200,11 +232,17 @@ class RestTransport_FpgFile(Resource):
             else:
                 filepath = request.args.get('fpgfile_path')
 
-            transportTarget.upload_to_ram_and_program(filepath)
-            response = {
-                'response':  transportTarget.fpgfile_path
-            }
-            code = 200
+            if not transportTarget.upload_to_ram_and_program(filepath):
+                response = {
+                    'response': ('Provided fpg file does not match '
+                        '`pr_template`: "{}"'.format(transportTarget.fpg_template))
+                }
+                code = 500
+            else:
+                response = {
+                    'response':  transportTarget.fpgfile_path
+                }
+                code = 200
         except:
             if filepath is not None:
                 response = {
@@ -224,8 +262,7 @@ class RestTransport_IsConnected(Resource):
         timeout = request.args.get('timeout', default = None, type = int)
         retries = request.args.get('retries', default = None, type = int)       
 
-        instance_id = getInstanceIdFromRequestArgs(request.args)        
-        transportTarget = getTransportTarget(target, instance_id)
+        transportTarget = getTransportTarget(target)
 
         try:
             response = {
@@ -241,8 +278,7 @@ class RestTransport_IsConnected(Resource):
 
 class RestTransport_IsProgrammed(Resource):
     def get(self, target):
-        instance_id = getInstanceIdFromRequestArgs(request.args)        
-        transportTarget = getTransportTarget(target, instance_id)
+        transportTarget = getTransportTarget(target)
 
         response = {
             'response':  transportTarget.fpgfile_path is not None
@@ -253,13 +289,13 @@ class RestTransport_IsProgrammed(Resource):
 class RestTransport_PciXdmaMap(Resource):
     def get(self):
         return {
-            'pci_xdma_id_map': XDMA_PCIE_DICT
+            'pci_xdma_id_map': PCIE_XDMA_DICT
         }, 200
 
 class RestTransport_Version(Resource):
     def get(self):
         return {
-            'response': '1.1.0'
+            'response': '1.2.0'
         }, 200
 
 api.add_resource(RestTransport_Device, '/<string:target>/device/<string:device_name>')
@@ -270,18 +306,42 @@ api.add_resource(RestTransport_IsProgrammed, '/<string:target>/programmed')
 api.add_resource(RestTransport_PciXdmaMap, '/PciXdmaMap')
 api.add_resource(RestTransport_Version, '/version')
 
-def get_xdma_pcie_map():
+def get_pcie_xdma_map():
     pcie_xdma_regex = r'/sys/bus/pci/drivers/xdma/\d+:(?P<pci_id>.*?):.*?/xdma/xdma(?P<xdma_id>\d+)_user'
     xdma_dev_filepaths = glob.glob('/sys/bus/pci/drivers/xdma/*/xdma/xdma*_user')
     ret = {}
     for fp in xdma_dev_filepaths:
         match = re.match(pcie_xdma_regex, fp)
-        ret[match.group('pci_id')] = int(match.group('xdma_id'))
+        ret[match.group('pci_id')] = match.group('xdma_id')
     return ret
 
-if __name__ == '__main__':
-    XDMA_PCIE_DICT = get_xdma_pcie_map()
-    app.run(host='0.0.0.0', port=5000, debug=False)  # run our Flask app
+if __name__ == '__main__':    
+    parser = argparse.ArgumentParser(
+        description=('Initialize a REST server exposing localpcietransports as'
+        'remote ones.')
+    )
+    parser.add_argument('fpgfile', type=str,
+        default='/home/cosmic/dev/vla-dev/pr_templates/adm_4x100g_pr_template_test/outputs/adm_4x100g_pr_template_test_2022-02-03_1951.fpg',
+        help='Path to the initial fpg file that the PCI devices are programmed with.'
+    )
+    parser.add_argument('--program', '-p', action='store_true',
+        help='Program the PCI devices with fpgfile.'
+    )
+    args = parser.parse_args()
+
+    PCIE_XDMA_DICT = get_pcie_xdma_map()
+    XDMA_PCIE_DICT = {}
+
+    for (pci_id, xdma_id) in PCIE_XDMA_DICT.items():
+        XDMA_PCIE_DICT[xdma_id] = pci_id
+        localtransport = getTransportTarget(xdma_id)
+        if args.program:
+            print('Programmed "pcie{}" successfully: {}'.format(pci_id, localtransport.upload_to_ram_and_program(args.fpgfile)))
+        else:
+            localtransport._setFpgfile_path(args.fpgfile)
+
+
+    app.run(host='0.0.0.0', port=5001, debug=False)  # run our Flask app
     # TODO disconnect instances when closing
     # for (target, cfpga) in TARGET_CFPGA_DICT.items():
     #     print('Disconnecting from "{}"'.format(target))
